Stop using Models.VPNConfiguration for VPN sessions
Models.VPNConfiguration was designed to describe user's selection of
authenticating instance, connecting instance and connecting profile.
This made it suitable for a data holder where all those three parameters
were required.

After time, it also became a data holder for the list of recently
selected profiles, adding the Popularity member.

When splitting instance and profile selection on the recent
configuration page the Models.VPNConfiguration has to disintegrate.
This patch is the first step towards that.