Force instance change notification on activating profile selection list
Navigating the client back and forth does not trigger the profile list
load if the connecting instance is not changed. Usually, that's not
required.

However, when the instance is broken and fails to return the profile
list the first time, navigating the client back and forth should try
to reload the profile list.