Cache failed authorization attempts
When the client was started or tried to connect to an instance which
access token expired, the authorization popup was displayed multiple
times - for each action thread, if authorization failed and/or was
canceled.

Furthermore, this commit upgrades the error message displayed on failed
authorization to make more sense to end-user unaware of OAuth tokens.

Signed-off-by: Simon Rozman <simon@rozman.si>