Revise forgetting logic
Forgetting the Secure Internet forgot to forget and reset the
authenticating instance.

Furthermore, forgetting logic was simplified to always forget what it
supposed to (at least what user expects). Regardless if the instance to
be forgotten is also a manually added part of another instance source.
User will need to re-authenticate in the later case.

Fixes: #130
Signed-off-by: Simon Rozman <simon@rozman.si>