Fix <MinisignPublicKey> iteration
XmlReader.ReadElementContentAsBase64() leaves the reader in a weird
state: it ends on the </MinisignPublicKey> as expected. But, the
following Read() or ReadEndElement() don't consume only the
</MinisignPublicKey>, but also the following element. In our case this
is the next <MinisignPublicKey> and the reader ends up on its text
content. This breaks the XML parsing, skipping every second
<MinisignPublicKey> entry.

XmlReader.ReadContentAsBase64() is no better.

Therefore, reading the <MinisignPublicKey>-s is manual now.

Signed-off-by: Simon Rozman <simon@rozman.si>