Stop reusing product GUID when MSI packages don't change
Using the same product GUID and package version, but different checksum
(due to recompilation) confuses the WiX bundle installer cache resulting
in uninstall of all "unchanged" packages on update.

The problem was not observed on x64 platform and was x86 specific due to
the order the packages of various platforms were processed.

Fixes: #140
Signed-off-by: Simon Rozman <simon@rozman.si>