Fix systray notifications when window is minimized
With the introduction of resizable and minimizable connection wizard
window, the visibility of the window is not defined only by the
IsVisible member any more.

The systray notifications are now displayed as soon as the connection
wizard window is not the active window any more - user is looking at
and/or working in another application.

Signed-off-by: Simon Rozman <simon@rozman.si>

a