Import Spanish translations from Transifex
Spanish (Spain) translation is not available for the time being and is
copied from English to prevent Latin American translations to appear in
Spain.

Unfortunately, the support for es-419 is lacking in .NET and MSI. The
.NET can be satisfied by declaring Latin American translations as
generic Spanish and overriding es-ES only. MSI is a bit more stubborn
and requires listing of all LCIDs of Latin American states to pick the
translation.

Translated-by: Carlos A. Pedreros Lizama <carlos.pedreros@gmail.com>
Signed-off-by: Simon Rozman <simon@rozman.si>