Clean institute access and secure internet settings when no discovery
On user settings migration from v1 to v2, all institute access and own
servers are copied into both lists: institute access and own servers.
Later when we get discovery information the lists are cleaned
appropriately.

However, in Let's Connect! client the discovery is disabled and the
lists are never cleaned. This prevents LC to return to search page after
the last server is forgotten.

Fixes: #164
Reported-by: François Kooman <fkooman@tuxed.net>
Signed-off-by: Simon Rozman <simon@rozman.si>