Installer: Switch to "uninstall before install" updating
We have been using "overwrite then cleanup" approach when updating.

However, with the Wintun 0.14 introduction, we need to cleanup previous
Wintun adapters. The 0.14 is using ephemeral adapters and it will not
reuse previously created adapters. Rather than introducing a big chunk
of cleanup code, we let previous version of eduVPN uninstall first, as
it already contains all the code to cleanup its adapters.

Signed-off-by: Simon Rozman <simon@rozman.si>