Simplify exception interception
As long as we're on the Window.Dispatcher thread the
Dispatcher.UnhandledException intercepts the exception and our event
implementation sets it to Window.Error for us.

When we're not on the Window.Dispatcher thread, it's enough to use
Window.Dispatcher.Invoke to rethrow the exception on the
Window.Dispatcher thread, then see the previous paragraph.

Signed-off-by: Simon Rozman <simon@rozman.si>