Connect to OpenVPN management interface from the client and not reverse
This discontinues --management-client. There was a window of opportunity
for an attacker to connect to our management interface socket and get
the management password from our client. Replacing our client on the
same socket before openvpn.exe is launched, would fool openvpn.exe to
trust it.

By using the original order - openvpn.exe listens on the management
interface socket, our client connects to it - an attacker would need to
guess the random generated session password in order to gain
openvpn.exe's trust.

Signed-off-by: Simon Rozman <simon@rozman.si>