Retry connecting to openvpn.exe when connection is refused
On really slow computers, openvpn.exe does not come up fast enough.
After openvpn.exe is launched by the Interactive Service and before
openvpn.exe manages to bind and listen on a communication port, nobody
is listening on the communication port yet and our client understood
this as a fatal error.

Reported-by: Rogier Spoor <Rogier.Spoor@SURFnet.nl>
Signed-off-by: Simon Rozman <simon@rozman.si>