Schedule user settings cleanup next time user logs on
We are using Active Setup because eduVPN client is installed per-machine
and eduVPN client settings are stored per-user. Active Setup allows us
to cleanup the settings for all users, not just the one that performed
the uninstall.

On an eduVPN client uninstall, Active Setup is instructed to cleanup
user's client settings on next user logon. It does not cleanup the
settings immediately on uninstall for the user performing the uninstall,
as it would require a reliable detection to distinguish the ultimate
uninstall from an uninstall-before-next-version-install.

This means, if user wants to start the eduVPN experience anew, one needs
to uninstall the client, sign out&in or reboot, and reinstall the
client.

Fixes: #187
Fixes: #192
Signed-off-by: Simon Rozman <simon@rozman.si>