Mark WireGuard tunnel service for removal after started
Should something interfere with WireGuard-connected client (e.g. a
computer reset), the tunnel .conf.dpapi and the service remained
installed. On next computer boot, SCM started the tunnel service and
computer automatically reconnected. However, this caused issues later
when user attempted to "reconnect" using the client resulting in failure
and ultimately a connectivity loss.

Note that a running service may be "deleted". SCM marks it for deletion.
After the service is stopped or on next boot the service is deleted.
Which is exactly what we want.

The only leftover remains the the .conf.dpapi file. We're not waiting
for the tunnel service to report running and we must not delete the
.conf.dpapi file before that.

Reported-by: JsBergbau
Fixes: #207
Signed-off-by: Simon Rozman <simon@rozman.si>