Stop caching endpoint paths and profiles
Should eduVPN server change (e.g. after an upgrade or a reconfiguration)
all running Windows clients should be restarted, as they used to cache
endpoint paths and profile list in memory. This is not a very good UX.

Endpoint paths are still cached, but 1 minute TTL is introduced.

Profile list is no longer explicitly cached. However, note it is loaded
in the profile selection drop-down after user selects the server to
connect to. In other words: this means that UI caches profile list while
user is making a profile selection. A new profile list refresh was
introduced after user disconnects, to attempt to bring the profile
selection drop-down up-to-date whenever it gets displayed.

Fixes: #217
Signed-off-by: Simon Rozman <simon@rozman.si>