Make authorization page a popup page
Authorization page was implemented as a normal page with reinvented
popup page behavior with one sole purpose: to hide the "<" navigate back
button. However, our implementation triggered OnActivate() on the page
the authorization page returned to again. This caused the search page to
download organization list again, which was a clear waste of resources.

Signed-off-by: Simon Rozman <simon@rozman.si>