Decouple server base URIs and organization IDs
eduvpn-common requires client to specify Secure Internet server by
organization ID (which is not necessarily an URI) and the country code.
What used to be called Server.Base property is therefore not always
the connecting server base URI, or URI at all.

As the client no longer contacts server API, we don't actually need the
base URI other than to identify a server.

Signed-off-by: Simon Rozman <simon@rozman.si>