Practice caution when restoring window from the system tray
Not sure, where the issue this commit is supposed to fix appears, as the
user didn't save the call stack. According to the resources available on
the web, I assume it was the Window.Show() method.

Closing this issue for now until more data is available. Preferably, a
call stack to see where this happens.

Fixes: #232
Reference: https://github.com/emoacht/Monitorian/commit/861d43dc34ad47f1243f3770086d857fe83aeeec
Signed-off-by: Simon Rozman <simon@rozman.si>