AND keywords to generate search hits
When "uni" is one of the search keywords, search produces large number
of hits. This makes UI laggish.

OR was used for analogy with the web search engines. However,  this
might not have been the best choice. When searching the web, user is not
sure what exactly is looking for. In eduVPN case, user knows exactly
what university is looking for.

Signed-off-by: Simon Rozman <simon@rozman.si>