Distribute servers between Institute Access and Other Servers on import
eduvpn-common does not support adding an Institute Access server that
is not listed in discovery. Admins have no other means to automate own
server configuration of clients.

So, when importing servers from registry (or 3.x client), the servers
are distributed between Institute Access and Other Servers according to
their discoverability.

Closes: #247
Signed-off-by: Simon Rozman <simon@rozman.si>