Revise failover test
1. Accessing Session.TunnelAddress from a failover thread should use
Wizard.TryInvoke to exclude any possible race condition.

2. Should the failover test fail re-fetch Session.TunnelAddress. OpenVPN
is known to report tunnel IP address later after already connected.
Thou, we're not using the failover test on OpenVPN connection, don't
assume WireGuard tunnel address is constant. Althou it always is because
it is a part of wg-conf.

3. Add try/catch outside failover test loop, should anything explode
there, the entire process would be killed.

Signed-off-by: Simon Rozman <simon@rozman.si>