Self-update: Sanitize Windows version in User-Agent header
This stop propagating Windows version numbering confusion. On the HTTP
server side regular expressions are typically used to parse the version,
so it is better to group all Windows builds of the same group and do the
regexp matching rather than performing numerical comparisons.

Mind that Windows 10 is quite a wide span range. Some features are
supported only from the specific build onward. I wouldn't make a big
deal - if the Go runtime, we will be switching to, requires a specific
Windows 10 build or newer, we'd simply push the update to all Windows 10
computers. Users reporting the client doesn't work on their Windows
should be instructed to update their Windows 10. It's free and no known
limitations, so why not.

Signed-off-by: Simon Rozman <simon@rozman.si>